requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
py-fibonacci==0.5.2
pycparser==2.22
PySocks==1.7.1
pytest==8.2.1
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-order==1.2.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
requests==2.32.2
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.12
tenacity==8.3.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
wsproto==1.2.0
